+++ packages/components/src/Utils/callback.ts
@@ -2,7 +2,7 @@ import { v4 as uuidv4 } from 'uuid';
 
 import { IContext } from './types';
 
-export const JsEvent = (context: IContext, Event: string, item: object = {}): Promise<unknown> => {
+export const JsEvent = (Event: string, item: object = {}, context: IContext = window.Alteryx): Promise<unknown> => {
   const callbackRegistrationId = uuidv4();
   const callbackPromise = new Promise(res => {
     context.Gui.Callbacks[callbackRegistrationId] = (...theArgs) => {