+++ packages/docs/components/DesignerApi/basicDemo.js
@@ -9,9 +9,9 @@ const DesignerApiDemo = () => {
     const [model, handleUpdateModel] = React.useContext(UiSdkContext);
 
     const incrementCount = () => {
-      let { count } = model.Configuration;
-      count++;
-      handleUpdateModel({ Configuration: { count } });
+      const newModel = { ...model }; 
+      newModel.Configuration.count++;
+      handleUpdateModel(newModel);
     };
     return (
       <AyxAppWrapper>