+++ packages/docs/releases/changelog.md
@@ -1,3 +1,10 @@
+### GA Date
+
+#### @alteryx/react-comms@1.0.0
+
+- Added support viewing the model in the CEF window developer tools. 
+- Updated JsEvent signature. Reordered the first parameter (context) to be the third and made it optional. 
+
 ### January 29th, 2021
 #### @ayx/ayx-ui-sdk@0.0.7-alpha